refactor: move Metal VLAN resource and data source to isolated package
