+++ equinix/provider.go
@@ -9,6 +9,7 @@ import (
 	"github.com/equinix/terraform-provider-equinix/internal/resources/metal/metal_connection"
 	"github.com/equinix/terraform-provider-equinix/internal/resources/metal/metal_project_ssh_key"
 	"github.com/equinix/terraform-provider-equinix/internal/resources/metal/metal_ssh_key"
+	metal_vlan "github.com/equinix/terraform-provider-equinix/internal/resources/metal/vlan"
 
 	"github.com/equinix/ecx-go/v2"
 	"github.com/equinix/terraform-provider-equinix/internal/config"
@@ -112,7 +113,7 @@ func Provider() *schema.Provider {
 			"equinix_metal_reserved_ip_block":    dataSourceMetalReservedIPBlock(),
 			"equinix_metal_spot_market_request":  dataSourceMetalSpotMarketRequest(),
 			"equinix_metal_virtual_circuit":      dataSourceMetalVirtualCircuit(),
-			"equinix_metal_vlan":                 dataSourceMetalVlan(),
+			"equinix_metal_vlan":                 metal_vlan.DataSource(),
 			"equinix_metal_vrf":                  dataSourceMetalVRF(),
 		},
 		ResourcesMap: map[string]*schema.Resource{
@@ -144,7 +145,7 @@ func Provider() *schema.Provider {
 			"equinix_metal_reserved_ip_block":    resourceMetalReservedIPBlock(),
 			"equinix_metal_ip_attachment":        resourceMetalIPAttachment(),
 			"equinix_metal_spot_market_request":  resourceMetalSpotMarketRequest(),
-			"equinix_metal_vlan":                 resourceMetalVlan(),
+			"equinix_metal_vlan":                 metal_vlan.Resource(),
 			"equinix_metal_virtual_circuit":      resourceMetalVirtualCircuit(),
 			"equinix_metal_vrf":                  resourceMetalVRF(),
 			"equinix_metal_bgp_session":          resourceMetalBGPSession(),
