+++ internal/resources/metal/vlan/resource_test.go
@@ -1,10 +1,11 @@
-package equinix
+package vlan_test
 
 import (
 	"fmt"
 	"log"
 	"testing"
 
+	"github.com/equinix/terraform-provider-equinix/internal/acceptance"
 	"github.com/equinix/terraform-provider-equinix/internal/config"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/acctest"
@@ -23,7 +24,7 @@ func init() {
 
 func testSweepVlans(region string) error {
 	log.Printf("[DEBUG] Sweeping vlans")
-	config, err := sharedConfigForRegion(region)
+	config, err := acceptance.GetConfigForNonStandardMetalTest()
 	if err != nil {
 		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting configuration for sweeping vlans: %s", err)
 	}
@@ -34,7 +35,7 @@ func testSweepVlans(region string) error {
 	}
 	pids := []string{}
 	for _, p := range ps {
-		if isSweepableTestResource(p.Name) {
+		if acceptance.IsSweepableTestResource(p.Name) {
 			pids = append(pids, p.ID)
 		}
 	}
@@ -46,7 +47,7 @@ func testSweepVlans(region string) error {
 			continue
 		}
 		for _, d := range ds.VirtualNetworks {
-			if isSweepableTestResource(d.Description) {
+			if acceptance.IsSweepableTestResource(d.Description) {
 				dids = append(dids, d.ID)
 			}
 		}
@@ -82,9 +83,9 @@ func TestAccMetalVlan_metro(t *testing.T) {
 	metro := "sv"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t) },
-		ExternalProviders: testExternalProviders,
-		Providers:         testAccProviders,
+		PreCheck:          func() { acceptance.TestAccPreCheckMetal(t) },
+		ExternalProviders: acceptance.TestExternalProviders,
+		Providers:         acceptance.TestAccProviders,
 		CheckDestroy:      testAccMetalVlanCheckDestroyed,
 		Steps: []resource.TestStep{
 			{
@@ -106,9 +107,9 @@ func TestAccMetalVlan_basic(t *testing.T) {
 	fac := "ny5"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t) },
-		ExternalProviders: testExternalProviders,
-		Providers:         testAccProviders,
+		PreCheck:          func() { acceptance.TestAccPreCheckMetal(t) },
+		ExternalProviders: acceptance.TestExternalProviders,
+		Providers:         acceptance.TestAccProviders,
 		CheckDestroy:      testAccMetalVlanCheckDestroyed,
 		Steps: []resource.TestStep{
 			{
@@ -135,7 +136,7 @@ func testAccCheckMetalVlanExists(n string, vlan *packngo.VirtualNetwork) resourc
 			return fmt.Errorf("No Record ID is set")
 		}
 
-		client := testAccProvider.Meta().(*config.Config).Metal
+		client := acceptance.TestAccProvider.Meta().(*config.Config).Metal
 
 		foundVlan, _, err := client.ProjectVirtualNetworks.Get(rs.Primary.ID, nil)
 		if err != nil {
@@ -152,7 +153,7 @@ func testAccCheckMetalVlanExists(n string, vlan *packngo.VirtualNetwork) resourc
 }
 
 func testAccMetalVlanCheckDestroyed(s *terraform.State) error {
-	client := testAccProvider.Meta().(*config.Config).Metal
+	client := acceptance.TestAccProvider.Meta().(*config.Config).Metal
 
 	for _, rs := range s.RootModule().Resources {
 		if rs.Type != "equinix_metal_vlan" {
@@ -185,9 +186,9 @@ func TestAccMetalVlan_importBasic(t *testing.T) {
 	fac := "ny5"
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t) },
-		ExternalProviders: testExternalProviders,
-		Providers:         testAccProviders,
+		PreCheck:          func() { acceptance.TestAccPreCheckMetal(t) },
+		ExternalProviders: acceptance.TestExternalProviders,
+		Providers:         acceptance.TestAccProviders,
 		CheckDestroy:      testAccMetalVlanCheckDestroyed,
 		Steps: []resource.TestStep{
 			{
